requirements.txt
fastapi
uvicorn
requests
matplotlib
nltk
yake
openai
python-dotenv
pydantic